feat: 메인 페이지에서 구글 로그인 동작하게 변경
